29 August

+++ hw08_mongo_library/composer/test/down.sh
@@ -1,1 +1,5 @@
-docker-compose down
+read -p "Are you sure(Y/n)? " -n 1 -r
+echo # (optional) move to a new line
+if [[ $REPLY =~ ^[YyДд]$ ]]; then
+    docker-compose down
+fi
